+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.46 ] - 2023/12/28
+
+  * Added method `SoundTouchClient`.`ToggleZoneMember` - Toggles the given zone member in the master device's zone.  If the member exists in the zone then it is removed; if the member does not exist in the zone, then it is added.  A new zone is automatically created if necessary.
+
 ###### [ 1.0.45 ] - 2023/12/26
 
   * Added `resolveSourceTitles:bool` argument to `SoundTouchClient`.`GetPresetList` and `GetRecentList` methods, so that a friendly source title can be displayed in user-interfaces.